++ itoa & strnew, should be working as intended

+++ ft_strnew.c
@@ -6,16 +6,17 @@
 /*   By: elehtora <elehtora@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/11/11 18:31:06 by elehtora          #+#    #+#             */
-/*   Updated: 2021/11/11 19:02:09 by elehtora         ###   ########.fr       */
+/*   Updated: 2021/11/27 02:56:26 by elehtora         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include <stdlib.h>
+#include <string.h>
 
 void	*ft_strnew(size_t size)
 {
 	char	*area;
-	int		i;
+	size_t	i;
 
 	area = (char *)malloc(size);
 	if (area == NULL)
